Hey,
Again, a nightly commit :) The last one because tomorrow I'll be in bed at a reasonable time since the morning after school starts... I wish they had a pill for _that_ ;) I think it's ready for release. Added more stability concerning the navigation box. It should now handle all possible URL's on a drupal site. To make sure I catch them all I implemented something that check whether a value is not set to one. It is set to one when the URL matches an option in the list. I saw that we also have a ?mod=moderation (which should only be visible to people that are logged in so I couldn't put it in the list standard) and for the rest nodes. Therefore, when in moderation/a node page it displays Moderation/Node in the list, else not. If I didn't do this Main would be set to the default value... we had that problem before... I needed to use $REQUEST_URI for the modules because it gives you the whole URL, $PHP_SELF only gives you a string until the first ? in the URL... This way I couldn't make up whether you were on the diary or book page. (module.php?mod=book and module.php?mod=diary) With a substr($REQUEST_URI,12,9) I could. Dries told me today that $REQUEST_URI isn't as good as $PHP_SELF, but this is an easy way to fix it. Voila I think that's it. (now that I think of it, I don't even have to do a substr() for the book module since there isn't anything that can be put behind it. With diary for example I have to do the substring because things like &op=add can be put behind it, but that's not the case with the book...) After this change I think that everything is fixed. Oh yes I also fixed a rather flagrant allignment bug. I align the big right td centered which would cause the comment's tree to be centered too so there wasn't much left of a tree anymore. ;) Added one table that fixed this. OK, I think the theme is ready for release. The only problem still is the vertical spacer which is fat in NS 4.76. If someone knows why or finds any other bugs in my theme please let me know. But for me, it's finished, until the next block code is stable and released, I'm planning to get rid of the account box too and expand the navigation menu somewhat. with indented options etc... but that's for later... Now I am going to sleep (after some more fixes I already have in mind) Bye!

Jeroen.

(I'm not sending the mail right away, first the fixes.) OK, it's 04.32. I did some other nice thing IMO, but it's not going to work anymore one we have more than 1000 nodes... Ack.. Just click a story or comment somewhere and check out the navigation box... Goodmorning. ;)