- Fixed bug 4710: comment module should use drupal_set_message() correctly.
- Fixed the title and breadcrumb trail of the comment delete page, added
missing cache flush, improved usability and made it so that when you delete
a comment replies are deleted as well.

- Added 'delete comment' links to the comments on node pages.  Hopefully
this will make it easier to maintain old threads such as those attached to
the book pages on drupal.org.

- Comment module maintenance: added missing t() functions and improved
consistency of the calls to drupal_set_message(), used radio buttons instead
of a selection box.